Добавляет PointController
